#65 GamesQuestions methods realocated to GamesQuestionsCtrl.
